Clone, blur, gradient and shapes, and a C long is 32 bits here
The four tools that need no kernel. Clone and blur join the brush rather than
sitting beside it: every mode differs only in what it puts down through the same
coverage — the settings' colour, a sample from somewhere else, or the tile
softened — so they share the tiles, the spacing and the opacity cap. Blur makes
its softened copy from a margin that reaches outside the tile, or the softening
would stop at a tile edge and leave a seam.

Shapes are built as outlines and filled by the selection rasteriser, so a circle
drawn with the shape tool and one selected with the ellipse marquee agree on
where the circle is. Gradients go through the same fill, which is why a gradient
against a selection has the selection's own edge rather than an approximation.

And a bug the tests caught the moment the kernels were called for real:
heal_coverage_bounds writes four C longs, which are 32 bits wide under the
Windows data model, so reading them as 64-bit integers took two of the kernel's
numbers for one. wand_mask's count had the same problem waiting.

+++ src/Compositor_korean_win.Core/Kernels.cs
@@ -73,8 +73,14 @@ public static partial class Kernels
     /// Marks every pixel within <paramref name="tolerance"/> of the one under the seed. Returns how
     /// many were selected, or -1 when memory runs out.
     /// </summary>
+    /// <remarks>
+    /// The kernel says <c>long</c>, which is 32 bits under the Windows data model however wide the
+    /// machine is — so this is an <c>int</c> here, and the same goes for
+    /// <see cref="HealCoverageBounds"/>'s four values. Declaring either as a 64-bit integer reads
+    /// two of the kernel's numbers as one.
+    /// </remarks>
     [LibraryImport(Library, EntryPoint = "wand_mask")]
-    public static partial long WandMask(nint rgba, nuint width, nuint height, nuint stride,
+    public static partial int WandMask(nint rgba, nuint width, nuint height, nuint stride,
                                         nuint seedX, nuint seedY, nuint radius, int tolerance,
                                         int contiguous, nint mask);
 
@@ -91,7 +97,7 @@ public static partial class Kernels
                                         out nint points, out nuint pointCount,
                                         out nint loops, out nuint loopCount);
 
-    /// <summary>Half-open bounds of nonzero bytes in a grey bitmap, as four longs.</summary>
+    /// <summary>Half-open bounds of nonzero bytes in a grey bitmap, as four 32-bit values.</summary>
     [LibraryImport(Library, EntryPoint = "heal_coverage_bounds")]
     public static partial void HealCoverageBounds(nint gray, nuint width, nuint height, nuint stride, nint bounds);
